Custom-App/requirements.txt
redis==6.2.0
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
psycopg2-binary==2.9.9
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
websocket-client==1.8.0
matplotlib==3.9.2
kiteconnect==5.0.1
python-dotenv==0.20.0
yfinance==0.2.61
sqlalchemy==1.4.49
openpyxl==3.1.5
reportlab==4.4.4
TA-Lib==0.6.8
python-multipart==0.0.20
prophet==1.1.5
cmdstanpy==1.2.0
mftool>=0.2.0
vaderSentiment==3.3.2
lightningchart==2.0.1
stocksent==0.4.5
textblob==0.19.0